Return ifo business climate data

Return ifo business climate data

Usage

ifo_business(
  type = c("germany", "sectors", "eastern", "saxony"),
  long_format = TRUE
)

Source

https://www.ifo.de/en/ifo-time-series

Arguments

type

character(1) Defaults to "germany". One of:

"germany": returns the ifo business climate index for Germany.

"sectors": returns the ifo business climate index for different sectors.

"eastern": returns the ifo business climate index for eastern Germany.

"saxony": returns the ifo business climate index for Saxony.

long_format

logical(1) if TRUE return the data in long format. Only applies to type "germany" and "sectors". Default TRUE.

Value

A data.frame() containing the monthly ifo business climate time series.

See also

Examples

# \donttest{
ifo_business("germany")
#> # A tibble: 1,398 × 6
#>    yearmonth  uncertainty economic_expansion indicator   series  value
#>    <date>           <dbl>              <dbl> <chr>       <chr>   <dbl>
#>  1 2005-01-01          NA               83.1 climate     index    92.2
#>  2 2005-01-01          NA               83.1 situation   index    87.4
#>  3 2005-01-01          NA               83.1 expectation index    97.2
#>  4 2005-01-01          NA               83.1 climate     balance   1.5
#>  5 2005-01-01          NA               83.1 situation   balance  -0.8
#>  6 2005-01-01          NA               83.1 expectation balance   3.8
#>  7 2005-02-01          NA               50.4 climate     index    92  
#>  8 2005-02-01          NA               50.4 situation   index    88  
#>  9 2005-02-01          NA               50.4 expectation index    96.2
#> 10 2005-02-01          NA               50.4 climate     balance   1  
#> # ℹ 1,388 more rows
# }